market/requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
autopep8==2.3.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.6.4
cryptography==43.0.3
defusedxml==0.8.0rc2
Django==5.1.2
django-baton==4.1.0
django-cachalot==2.6.3
django-environ==0.11.2
django-extensions==3.2.3
django-redis==5.4.0
django-silk==5.2.0
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
gprof2dot==2024.6.6
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
MarkupSafe==3.0.2
oauthlib==3.2.2
packaging==24.1
pluggy==1.5.0
prompt_toolkit==3.0.48
pycodestyle==2.12.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
pytest==8.3.3
pytest-cov==5.0.0
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.2
redis==5.2.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.0
sentry-sdk==2.17.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.6
